+++ src/Application.php
@@ -12,6 +12,8 @@ use Chassis\Framework\Brokers\Amqp\Configurations\BrokerConfigurationInterface;
 use Chassis\Framework\Brokers\Amqp\Contracts\ContractsManager;
 use Chassis\Framework\Brokers\Amqp\Contracts\ContractsManagerInterface;
 use Chassis\Framework\Brokers\Amqp\Contracts\ContractsValidator;
+use Chassis\Framework\Brokers\Amqp\Handlers\MessageHandler;
+use Chassis\Framework\Brokers\Amqp\Handlers\MessageHandlerInterface;
 use Chassis\Framework\Brokers\Amqp\Streamers\PublisherStreamer;
 use Chassis\Framework\Brokers\Amqp\Streamers\PublisherStreamerInterface;
 use Chassis\Framework\Brokers\Amqp\Streamers\SubscriberStreamer;
@@ -225,25 +227,16 @@ class Application extends Container implements ArrayAccess
         // container bindings
         $this->add(ContractsManagerInterface::class, ContractsManager::class)
             ->addArguments([BrokerConfigurationInterface::class, new ContractsValidator()]);
-        $this->add('brokerStreamConnection', function ($app) {
+        $this->add('brokerStreamConnection', function ($contractsManager) {
             return new AMQPStreamConnection(
-                ...array_values(
-                    $app->get(ContractsManagerInterface::class)->toStreamConnectionFunctionArguments()
-                )
+                ...array_values($contractsManager->toStreamConnectionFunctionArguments())
             );
-        })->addArgument($this)->setShared(false);
+        })->addArgument(ContractsManagerInterface::class);
+        $this->add(MessageHandlerInterface::class, MessageHandler::class);
         $this->add(SubscriberStreamerInterface::class, SubscriberStreamer::class)
-            ->addArguments([
-                $this->get('brokerStreamConnection'),
-                ContractsManagerInterface::class,
-                LoggerInterface::class
-            ])->setShared(false);
+            ->addArgument($this)->setShared(false);
         $this->add(PublisherStreamerInterface::class, PublisherStreamer::class)
-            ->addArguments([
-                $this->get('brokerStreamConnection'),
-                ContractsManagerInterface::class,
-                LoggerInterface::class
-            ])->setShared(false);
+            ->addArgument($this)->setShared(false);
     }
 
     /**