+++ src/Framework/Services/BrokerAbstractService.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Chassis\Framework\Services;
 
 use Chassis\Application;
+use Chassis\Framework\Brokers\Amqp\BrokerRequest;
 use Chassis\Framework\Brokers\Amqp\BrokerResponse;
 use Chassis\Framework\Brokers\Amqp\MessageBags\MessageBagInterface;
 
@@ -14,21 +15,31 @@ class BrokerAbstractService implements ServiceInterface
 {
     protected const LOGGER_COMPONENT_PREFIX = "broker_service_";
 
-    protected MessageBagInterface $messageBag;
+    protected MessageBagInterface $message;
     protected Application $app;
 
     /**
-     * @param MessageBagInterface $messageBag
+     * @param MessageBagInterface $message
+     *
+     * @throws \Psr\Container\ContainerExceptionInterface
+     * @throws \Psr\Container\NotFoundExceptionInterface
      */
     public function __construct(
-        MessageBagInterface $messageBag
+        MessageBagInterface $message
     ) {
-        $this->messageBag = $messageBag;
+        $this->message = $message;
         $this->app = app();
     }
 
     public function response($body = []): BrokerResponse
     {
-        return new BrokerResponse($body, []);
+        return (new BrokerResponse($body, []))
+            ->fromContext($this->message);
+    }
+
+    public function request(string $operation, array $body): BrokerRequest
+    {
+        return (new BrokerRequest($body))
+            ->fromContext($this->message, $operation);
     }
 }