+++ src/Framework/Workers/Worker.php
@@ -4,42 +4,38 @@ declare(strict_types=1);
 
 namespace Chassis\Framework\Workers;
 
-use Chassis\Framework\Brokers\Amqp\Configurations\BrokerConfigurationInterface;
-use Chassis\Framework\Brokers\Amqp\Contracts\ContractsManager;
-use Chassis\Framework\Brokers\Amqp\Contracts\ContractsValidator;
+use Chassis\Application;
 use Chassis\Framework\Brokers\Amqp\Streamers\InfrastructureStreamer;
 use Chassis\Framework\Brokers\Amqp\Streamers\SubscriberStreamer;
+use Chassis\Framework\Brokers\Exceptions\StreamerChannelIterateMaxRetryException;
 use Chassis\Framework\InterProcessCommunication\ChannelsInterface;
 use Chassis\Framework\InterProcessCommunication\DataTransferObject\IPCMessage;
 use Chassis\Framework\InterProcessCommunication\ParallelChannels;
-use Chassis\Framework\Routers\Router;
-use Chassis\Framework\Routers\RouterInterface;
 use Chassis\Framework\Threads\Exceptions\ThreadConfigurationException;
-use Psr\Log\LoggerInterface;
 use Throwable;
 
-use function Chassis\Helpers\app;
 use function Chassis\Helpers\subscribe;
 
 class Worker implements WorkerInterface
 {
     private const LOGGER_COMPONENT_PREFIX = "worker_";
-    private const LOOP_EACH_MS = 50;
+    private const SUBSCRIBER_ITERATE_MAX_RETRY = 5;
 
-    private SubscriberStreamer $subscriberStreamer;
+    private Application $application;
     private ChannelsInterface $channels;
-    private LoggerInterface $logger;
+    private SubscriberStreamer $subscriberStreamer;
+    private int $iterateRetry = 0;
 
     /**
+     * @param Application $application
      * @param ChannelsInterface $channels
-     * @param LoggerInterface $logger
      */
     public function __construct(
-        ChannelsInterface $channels,
-        LoggerInterface $logger
+        Application $application,
+        ChannelsInterface $channels
     ) {
+        $this->application = $application;
         $this->channels = $channels;
-        $this->logger = $logger;
     }
 
     /**
@@ -50,17 +46,16 @@ class Worker implements WorkerInterface
         try {
             $this->subscriberSetup();
             do {
-                $startAt = microtime(true);
-                // channel event poll & streamer iterate
+                // IPC channel event poll
                 if (!$this->polling()) {
                     break;
                 }
-                // Wait a while - prevent CPU load
-                $this->loopWait($startAt);
+                // subscriber streamer iterate
+                $this->subscriberIterate();
             } while (true);
         } catch (Throwable $reason) {
             // log this error & request respawning
-            $this->logger->error(
+            $this->application->logger()->error(
                 $reason->getMessage(),
                 [
                     'component' => self::LOGGER_COMPONENT_PREFIX . "exception",
@@ -72,30 +67,57 @@ class Worker implements WorkerInterface
                 (new IPCMessage())->set(ParallelChannels::METHOD_RESPAWN_REQUESTED)
             );
         }
+
+        // Close subscriber streamer channel
+        if (isset($this->subscriberStreamer)) {
+            $this->subscriberStreamer->closeChannel();
+        }
     }
 
     /**
      * @return bool
      */
-    private function polling(): bool
+    protected function polling(): bool
     {
         // channel events pool
-        $polling = $this->channels->eventsPoll();
-        if ($this->channels->isAbortRequested()) {
-            // send aborting message to main thread
-            $this->channels->sendTo(
-                $this->channels->getThreadChannel(),
-                (new IPCMessage())->set(ParallelChannels::METHOD_ABORTING)
-            );
-            return false;
+        $this->channels->eventsPoll();
+        if (!$this->channels->isAbortRequested()) {
+            return true;
         }
 
-        // subscriber iterate
-        if (isset($this->subscriberStreamer)) {
+        // send aborting message to thread manager
+        $this->channels->sendTo(
+            $this->channels->getThreadChannel(),
+            (new IPCMessage())->set(ParallelChannels::METHOD_ABORTING)
+        );
+
+        return false;
+    }
+
+    /**
+     * @return void
+     * @throws StreamerChannelIterateMaxRetryException
+     */
+    protected function subscriberIterate(): void
+    {
+        try {
+            if (!isset($this->subscriberStreamer)) {
+                // threads without subscriber need to wait more
+                usleep(500000);
+                return;
+            }
             $this->subscriberStreamer->iterate();
-        }
+            $this->iterateRetry = 0;
 
-        return $polling;
+        } catch (Throwable $reason) {
+            // retry pattern
+            $this->iterateRetry++;
+            if ($this->iterateRetry >= self::SUBSCRIBER_ITERATE_MAX_RETRY) {
+                throw new StreamerChannelIterateMaxRetryException("streamer channel iterate - to many retry");
+            }
+            // wait before retry
+            sleep(1);
+        }
     }
 
     /**
@@ -105,27 +127,18 @@ class Worker implements WorkerInterface
      */
     protected function subscriberSetup(): void
     {
-        $threadConfiguration = app('threadConfiguration');
+        $threadConfiguration = $this->application->get('threadConfiguration');
         switch ($threadConfiguration["threadType"]) {
             case "infrastructure":
                 // Broker channels setup
-                (new InfrastructureStreamer(
-                    app()->get('brokerStreamConnection'),
-                    new ContractsManager(
-                        app(BrokerConfigurationInterface::class),
-                        new ContractsValidator()
-                    ),
-                    $this->logger
-                ))->brokerChannelsSetup();
+                (new InfrastructureStreamer($this->application))->brokerChannelsSetup();
                 break;
             case "configuration":
-                // wait after infrastructure to finish exchange/queues/bindings declarations
-                usleep(rand(2000000, 4000000));
                 // TODO: implement configuration listener - (centralized configuration server feature)
                 break;
             case "worker":
-                // wait after infrastructure to finish exchange/queues/bindings declarations
-                usleep(rand(2000000, 5000000));
+                // wait a while - infrastructure must declare exchanges, queues & bindings
+                usleep(rand(2500000, 4000000));
                 // create subscriber
                 $this->subscriberStreamer = subscribe(
                     $threadConfiguration["channelName"],
@@ -134,19 +147,6 @@ class Worker implements WorkerInterface
                 break;
             default:
                 throw new ThreadConfigurationException("unknown thread type");
-        }
-    }
-
-    /**
-     * @param float $startAt
-     *
-     * @return void
-     */
-    private function loopWait(float $startAt): void
-    {
-        $loopWait = self::LOOP_EACH_MS - (round((microtime(true) - $startAt) * 1000));
-        if ($loopWait > 0) {
-            usleep(((int)$loopWait * 1000));
         }
     }
 }